updating

+++ zoodatasets/chunkdatasets.py
@@ -88,7 +88,7 @@ class ZooDataset(Dataset):
     def __getitem__(self, idx):
         if torch.is_tensor(idx):
             idx = idx.tolist()
-        weight = self.data[idx]
+        weight = self.data[idx].to(torch.float32)
         if self.transform:
             weight = self.transform(weight)
